+++ containers/App.js
@@ -1,16 +1,16 @@
-import { bindActionCreators } from 'redux';
-import { connect }            from 'react-redux';
-import * as CounterActions    from '../actions/counter';
+import React, { Component, PropTypes } from 'react';
+import Image                  from '../components/Image';
 import Counter                from '../components/Counter';
 
-function mapStateToProps(state) {
-  return {
-    counter: state.counter
-  };
+class App extends Component {
+  render() {
+    return(
+      <div>
+        <Counter />
+        <Image imageSrc={this.props.imageSrc} />
+      </div>
+    )
+  }
 }
 
-function mapDispatchToProps(dispatch) {
-  return bindActionCreators(CounterActions, dispatch);
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(Counter);
+export default App
